+++ brainscore/tolerance.py
@@ -5,9 +5,9 @@ from brainscore.benchmarks.majajhong2015_combined import _DicarloMajajHong2015Re
 from brainscore.metrics.ceiling import InternalConsistency, RDMConsistency, ToleranceConsistency
 from brainscore.metrics.rdm import RDMCrossValidated
 from brainscore.metrics.regression import CrossRegressedCorrelation, mask_regression, ScaledCrossRegressedCorrelation, \
-    pls_regression, gram_control_regression, old_gram_control_pls, pearsonr_correlation
+    pls_regression, pearsonr_correlation
 from brainscore.metrics.regression_extra import CrossRegressedCorrelationCovariate, semipartial_regression, \
-    semipartial_pls, gram_pls, \
+    semipartial_pls, gram_pls, gram_control_regression, old_gram_control_pls, \
     ToleranceCrossValidation, CrossRegressedCorrelationDrew
